F-15 migration to module
Quite many controls had either lowercase letter, one were duplicate, several had double underscore. Several with same identifier which were renamed. Memory addresses changed since several duplicates where renamed.
